moved character data transform functions into utils
